Clean up comments and file structure
